Revert "feat(topology): Add topology to dsp and migrate to v1beta2"
