+++ k8s/operators/src/pool/diskpool/mod.rs
@@ -3,9 +3,6 @@
 
 /// DiskPool client operations.
 pub(crate) mod client;
-/// DiskPool migration operations.
-pub(crate) mod migration;
 /// The DiskPool custom resource definition.
 pub(crate) mod v1alpha1;
 pub(crate) mod v1beta1;
-pub(crate) mod v1beta2;
